fix: simplify MathJS use
